Store CPU-side GPU programs in hash maps
This changeset prevents the renderer for loading the same (vertex
shader, fragment shader) pair multiple times.

+++ base/include/render/Material.hpp
@@ -23,7 +23,7 @@ namespace donkey {
 namespace render {
 
 struct Material : Resource {
-  std::uint32_t program_id;
+  std::size_t program_id;
   unsigned int position_location;
   unsigned int normal_location;
   unsigned int uv_location;
@@ -42,7 +42,7 @@ struct Material : Resource {
   unsigned int camera_position_location;
 
   Material(std::uint32_t id,
-           std::uint32_t program_id,
+           std::size_t program_id,
            int position_location,
            int normal_location,
            int uv_location,
